requirements.txt
bs4
lxml
cchardet
lyricsgenius
python-slugify